requirements.txt
langchain
langchain-groq
python-dotenv
streamlit
hugchat==0.0.8